:recycle: api: move SessionStartedDeletionError to src
