+++ api/lib/application/error-manager.js
@@ -251,9 +251,6 @@ function _mapToHttpError(error) {
   if (error instanceof DomainErrors.SessionWithAbortReasonOnCompletedCertificationCourseError) {
     return new HttpErrors.ConflictError(error.message, error.code, error.meta);
   }
-  if (error instanceof DomainErrors.SessionStartedDeletionError) {
-    return new HttpErrors.ConflictError(error.message);
-  }
   if (error instanceof DomainErrors.OrganizationLearnerAlreadyLinkedToUserError) {
     return new HttpErrors.ConflictError(error.message, error.code, error.meta);
   }
