+++ api/src/certification/session/domain/errors.js
@@ -1,0 +1,9 @@
+import { DomainError } from '../../../shared/domain/errors.js';
+
+class SessionStartedDeletionError extends DomainError {
+  constructor(message = 'La session a déjà commencé.') {
+    super(message);
+  }
+}
+
+export { SessionStartedDeletionError };
