+++ api/src/certification/session/domain/usecases/delete-session.js
@@ -1,4 +1,4 @@
-import { SessionStartedDeletionError } from '../../../../../lib/domain/errors.js';
+import { SessionStartedDeletionError } from '../errors.js';
 
 /**
  * @typedef {import ('../../../shared/domain/usecases/index.js').dependencies} deps
